검색 화면 아이콘 수정, VIew, ViewModel 수정완료.
