+++ package.json
@@ -15,6 +15,7 @@
     "@gorhom/bottom-sheet": "^5.2.6",
     "@react-native-async-storage/async-storage": "^2.1.1",
     "@react-native-clipboard/clipboard": "1.16.2",
+    "@react-native-community/blur": "^4.4.1",
     "@react-native-firebase/app": "21.12.0",
     "@react-native-google-signin/google-signin": "^13.2.0",
     "@react-native-seoul/kakao-login": "5.4.1",
@@ -26,6 +27,7 @@
     "@react-navigation/stack": "^7.1.1",
     "axios": "^1.8.4",
     "expo-font": "^13.0.4",
+    "expo-modules-core": "^3.0.24",
     "geolocation": "^0.2.0",
     "jwt-decode": "^4.0.0",
     "lodash.truncate": "^4.4.2",
